# base-api/openapi.yaml
openapi: 3.0.3
info:
  title: B.A.S.E. Identify / Prove API
  version: "1.0.0"
  description: |
    Evidence-driven HW/FW reverse engineering API.
    Canonical v1: identify · prove · usage.
    No LLM in the proof path. saas_production: false (lab meter).
  license:
    name: AGPL-3.0
servers:
  - url: http://127.0.0.1:8787
    description: Local lab
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
  /v1/identify:
    post:
      tags: [canonical]
      summary: Identify firmware/HW per contract
      operationId: identify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentifyRequest"
      responses:
        "200":
          description: Identification + usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  responses:
    Unauthorized:
      description: Missing or invalid API key
    PaymentRequired:
      description: Insufficient credits
    BadRequest:
      description: Invalid body
  schemas:
    IdentifyRequest:
      type: object
      required: [firmware_b64, mmio]
      properties:
        firmware_b64:
          type: string
          description: Base64 firmware bytes
        mmio:
          type: array
          items:
            $ref: "#/components/schemas/MmioAccess"
        contracts_yaml:
          type: string
          description: Optional SequenceContract YAML list
        label:
          type: string
    MmioAccess:
      type: object
      required: [address, access_type]
      properties:
        address:
          type: integer
          format: int64
        value:
          type: integer
          format: int64
          nullable: true
        access_type:
          type: string
          enum: [read, write]
        function_name:
          type: string
        instruction_addr:
          type: integer
          format: int64
    IdentifyResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: identify.result
        hardware:
          type: object
        by_contract:
          type: array
          items:
            type: object
        usage:
          $ref: "#/components/schemas/UsageBreakdown"
        honesty:
          type: object
...
    UsageBreakdown:
      type: object
      properties:
        firmware_bytes:
          type: integer
        mmio_events:
          type: integer
        contracts:
          type: integer
        units:
          type: integer
        credits_charged:
          type: integer
        credits_remaining:
          type: integer
...